requirements.txt
python-gitlab==4.4.0
PyYAML==6.0
psycopg2-binary==2.9.6
#jira==2.0.0 # for jira and may be outdated
#ipython # for jira and may be outdated
requests==2.31.0
deepdiff
rich
gql
docker
click
